requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
lxml==5.3.1
packaging==24.2
PyMuPDF==1.25.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-docx==1.1.2
python-dotenv==1.1.0
typing_extensions==4.13.0
